% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeefive.R
\name{geefive_pal}
\alias{geefive_pal}
\title{G5 color palette}
\usage{
geefive_pal()
}
\description{
The standard three-color G5 palette for line plots comprises
blue, blue-grey, vermillion
}
\seealso{
Other colour geefive: \code{\link{scale_colour_geefive}}
}
\concept{colour geefive}
