% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeefive.R
\name{theme_geefive}
\alias{theme_geefive}
\title{Theme inspired by G5 Branding}
\usage{
theme_geefive(base_size = 12, base_family = "mono")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme inspired by the plots on
\href{G5}{http://getg5.com}.
}
\concept{themes g5}
